python: move `_resolve_args_directness` function
Needed for the next commit.
